Split HTTP code into a separate module
